Remove incorrect version of total_seconds()
The second copy of total_seconds() is incorrect as it looks for
the total_seconds method in datetime.datetime, not datetime.timedelta.

convert_to_epoch() still worked since it would use the first version,
but any other user of total_seconds would get the second, incorrect
version which always takes the else branch.

This commit simply removes the incorrect version.